Restructured and added python cli
